backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
akshare==1.12.0
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
aiohttp==3.9.1
schedule==1.2.1
